Consolidate and fix readability for main transformer layer
